# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/categories/{categoryId}" : {
      "put" : {
        "tags" : [ "category-controller" ],
        "operationId" : "updateCategory",
        "parameters" : [ {
          "name" : "categoryId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CategoryDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCategoryDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/outlets/{outletId}/categories/timings" : {
      "post" : {
        "tags" : [ "category-controller" ],
        "operationId" : "setOutletCategoryTimings",
        "parameters" : [ {
          "name" : "outletId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SetOutletCategoryTimingRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListTimingDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/categories" : {
      "get" : {
        "tags" : [ "category-controller" ],
        "operationId" : "getCategories",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListCategoryDTO"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "category-controller" ],
        "operationId" : "createCategory",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CategoryDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCategoryDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/brands/{brandId}/categories" : {
      "get" : {
        "tags" : [ "category-controller" ],
        "operationId" : "getBrandCategories",
        "parameters" : [ {
          "name" : "brandId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListCategoryDTO"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "category-controller" ],
        "operationId" : "createBrandCategory",
        "parameters" : [ {
          "name" : "brandId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CategoryDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCategoryDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/brands/{brandId}/categories/timings" : {
      "post" : {
        "tags" : [ "category-controller" ],
        "operationId" : "setBrandCategoryTimings",
        "parameters" : [ {
          "name" : "brandId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SetBrandCategoryTimingRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListTimingDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/outlets/{outletId}/categories/{categoryId}/timings" : {
      "get" : {
        "tags" : [ "category-controller" ],
        "operationId" : "getOutletCategoryTimings",
        "parameters" : [ {
          "name" : "outletId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "categoryId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListTimingDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/brands/{brandId}/categories/{categoryId}/timings" : {
      "get" : {
        "tags" : [ "category-controller" ],
        "operationId" : "getBrandCategoryTimings",
        "parameters" : [ {
          "name" : "brandId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "categoryId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListTimingDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CategoryDTO" : {
        "required" : [ "name" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "brandId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "maxLength" : 100,
            "minLength" : 2,
            "type" : "string"
          },
          "description" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "timings" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CategoryTimingDTO"
            }
          }
        }
      },
      "CategoryTimingDTO" : {
        "required" : [ "closingTime", "openingTime" ],
        "type" : "object",
        "properties" : {
          "openingTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "closingTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          }
        }
      },
      "LocalTime" : {
        "type" : "object",
        "properties" : {
          "hour" : {
            "type" : "integer",
            "format" : "int32"
          },
          "minute" : {
            "type" : "integer",
            "format" : "int32"
          },
          "second" : {
            "type" : "integer",
            "format" : "int32"
          },
          "nano" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ApiResponseCategoryDTO" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/CategoryDTO"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "SetOutletCategoryTimingRequest" : {
        "required" : [ "categoryId", "timings" ],
        "type" : "object",
        "properties" : {
          "categoryId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "timings" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TimingDTO"
            }
          }
        }
      },
      "TimingDTO" : {
        "required" : [ "closingTime", "openingTime" ],
        "type" : "object",
        "properties" : {
          "openingTime" : {
            "type" : "string",
            "example" : "10:00:00"
          },
          "closingTime" : {
            "type" : "string",
            "example" : "22:00:00"
          }
        }
      },
      "ApiResponseListTimingDTO" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TimingDTO"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "SetBrandCategoryTimingRequest" : {
        "required" : [ "categoryId" ],
        "type" : "object",
        "properties" : {
          "categoryId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "timings" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TimingDTO"
            }
          }
        }
      },
...
      "ApiResponseListCategoryDTO" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CategoryDTO"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    }
  }
}